fix(agents): tighten orchestration contracts
